Get /archivematica/dashboard to redirect to a login page
Before this patch, trying to access the dashboard when not logged in
went as follows:

GET https://workflow.wc.org/archivematica/dashboard/
302 FOUND > workflow.wc.org/administration/accounts/login/

which 404s, because it's not an Archivematica specific path.

It now has the slightly more correct behaviour:

GET https://workflow.wc.org/archivematica/dashboard/
302 FOUND > workflow.wc.org/archivematica/dashboard/administration/accounts/login/

which returns the login page.

We do this by passing the SCRIPT_NAME env var to WSGI, which tells it
to prefix most URLs with /archivematica/dashboard (and silently removes
it from Django for us) -- it's just the hard-coded login in
middleware.py that causes us some annoyance.