Add makefile commands for bootstrapping db
Plus a couple of other useful entries

Remove the mysql-bootstrap command from docker-compose - it's helpful to
be able to run this synchronously so we know when the database has been
set up and we can restart services. (Docker will just fire up the
container and let it run, with seemingly no way to wait until it's
exited before running dependent commands).