Ensure we mount the permanent volume, don't recreate it every time
Previously we were creating a new EBS volume every time we booted an
instance, so any data written to it would be lost when the instance
rebooted.  Now we use the permanent volume!

The volume also lives in the critical stack, because the data it contains
is fairly important and should be persisted.