Look at more XML files in the transfer monitor
When the transfer monitor runs, it's looking for the METS file in the
reporting cluster to see that the transfer completed successfully.
Unfortunately, the name of that file is structured in a way that makes
it slightly difficult to query:

    data/objects/submissionDocumentation/transfer-ARTCOOB14-46f0a350-d18d-4123-a162-70589bf9786e/METS.xml

So it has to query the list of storage service files in an approximate
way:

    SELECT files
    WHERE externalIdentifier={archivematicaIdentifier}
    AND suffix=".xml"

which can fail if a bag contains lots of XML files and the METS file
isn't in the first page of results.

We can fix this in the storage service, but we haven't done that work yet.

This ups the result window from 100 to 10,000, so it's more likely we'll
find the METS file in that first page of results.  If this still isn't
enough, we can go back to the storage service.