Use hyphen instead of underscore in image name to account for updater docker version
This is a fix for https://github.com/buildkite-plugins/docker-compose-buildkite-plugin/issues/410
where the underlying docker compose version has changed, and the naming convention for docker
images has shifted from using an underscore to a hyphen.

We can fix this by updating our build scripts rather than by forcing the docker version of the plugin
as has been done elsewhere (https://github.com/wellcomecollection/content-api/pull/82) as we are not
using the docker-composer plugin.