Increase the memory available to the transfer_monitor_lambda
We've noticed this lambda failing through timeouts. Upon investigation it appeas the lambda has been conistently using its entire allocation of 128MB which suggested that it was insufficiciently resourced resulting in it locking up and timing out.

After experimentation is seems that 512MB is sufficient to cause the transfer lambda to succeed. This change will result in some increased cost but as this lambda runs quite infrequently the cost change is negligable.